Move pipelines.ts and root.ts to index.ts :(
